Feature/pay slip

+++ app/src/main/java/br/com/if91wallet/viewmodel/TransactionViewModel.kt
@@ -8,5 +8,5 @@ import br.com.if91wallet.vo.UserVo
 class TransactionViewModel (private val repository: TransactionRepository) : ViewModel() {
 
     fun getTransaction() = repository.transaction
-    fun pay(value: Double, card: CardVo, user: UserVo) = repository.pay(value, card, user)
+    fun pay(value: Float, card: CardVo, user: UserVo) = repository.pay(value, card, user)
 }
